fixed verify (I hope)

+++ packages/app/src/pages/api/apply/verify.ts
@@ -6,7 +6,7 @@ import {
   uploadFile,
   UploadFolder,
 } from 'lib/services/directus/server'
-import { sendNotificationEmail } from 'lib/services/sendgrid/server'
+import { SendGridCategory, SendGridTemplate, sendNotificationEmail } from 'lib/services/sendgrid/server'
 import { withApplicant, withMethods } from 'lib/utils/server'
 import { NextApiRequest, NextApiResponse } from 'next'
 
@@ -21,34 +21,38 @@ async function Verify(req: NextApiRequest, res: NextApiResponse<ApiResponse<Appl
       previousPhoto = applicant.photo as DirectusFile
     }
 
+    let file = null
     const fileInfo = await getFileInfo(req)
-    const file = await uploadFile(
-      fileInfo,
-      UploadFolder.verification,
-      `Verification: ${applicant.id.substring(0, 4)}-${applicant.id.substring(4, 8)}`,
-      `Verification for ${applicant.email}: ${applicant.first_name} ${applicant.last_name} `
-    )
+    if (fileInfo) {
+      file = await uploadFile(
+        fileInfo,
+        UploadFolder.verification,
+        `Verification: ${applicant.id.substring(0, 4)}-${applicant.id.substring(4, 8)}`,
+        `Verification for ${applicant.email}: ${applicant.first_name} ${applicant.last_name} `
+      )
+    }
 
     const updatedUser = (await updateUser(applicant.id, {
-      photo: file.id,
+      photo: file?.id || previousPhoto?.id || null,
       application_status: 'review',
       user_type: 'applicant',
     })) as Applicant
 
-    const status = ApplicationStatus[applicant.application_status]
-    if (status == ApplicationStatus.verify && previousPhoto == null)
-      await sendNotificationEmail(
-        applicant.email,
-        applicant.nickname || applicant.first_name + ' ' + applicant.last_name,
-        `Application Status`,
-        'Your photo ID was submitted. It may take a few days to review.',
-        {
-          button_text: 'Check Application Results',
-          button_url: 'https://guysnheat.com/apply',
-        }
-      )
 
-    if (previousPhoto) {
+    await sendNotificationEmail(
+      applicant.email,
+      applicant.nickname || applicant.first_name + ' ' + applicant.last_name,
+      `Application Status`,
+      'Your photo was submitted. It may take a few days to review.',
+      {
+        button_text: 'Check Application Results',
+        button_url: 'https://guysnheat.com/apply',
+      },
+      SendGridTemplate.AppNotification,
+      SendGridCategory.Notification
+    )
+
+    if (fileInfo && previousPhoto) {
       await deleteFile(previousPhoto.id)
     }
 
@@ -60,6 +64,7 @@ async function Verify(req: NextApiRequest, res: NextApiResponse<ApiResponse<Appl
 }
 
 export default Verify
+
 export const config = {
   api: {
     bodyParser: false,
